Refactoring inline editor and css inline editor to separate files
