+++ src/editor/InlineEditorProviders.js
@@ -18,6 +18,7 @@ define(function (require, exports, module) {
         DocumentManager     = require("document/DocumentManager"),
         EditorManager       = require("editor/EditorManager"),
         FileUtils           = require("file/FileUtils"),
+        CSSInlineEditor     = require("editor/CSSInlineEditor"),
         ProjectManager      = require("project/ProjectManager");
 
     // track divs to re-position manually
@@ -172,195 +173,7 @@ define(function (require, exports, module) {
         }, 0);
     }
     
-    /**
-     * @constructor
-     *
-     */
-    function InlineEditor() {
-    }
-    
-
-    InlineEditor.prototype.htmlContent = null;
-    InlineEditor.prototype.editor = null;
-    InlineEditor.prototype.height = 0;
-    InlineEditor.prototype.inlineId = null;
-    InlineEditor.prototype.hostEditor = null;
-
-    // TY TODO: do this another way?
-    // Update the inline editor's height when the number of lines change
-    InlineEditor.prototype.prevHeight = 0;
-
-    /**
-     * @private
-     * Given a host editor and its inline editors, find the widest gutter and make all the others match
-     * @param {!Editor} hostEditor Host editor containing all the inline editors to sync
-     */
-    InlineEditor.prototype.syncGutterWidths = function () {
-        var editors = EditorManager.getInlineEditors(this.hostEditor);
-        // add the host to the list and go through them all
-        editors.push(this.hostEditor);
         
-        var maxWidth = 0;
-        editors.forEach(function (editor) {
-            var gutter = $(editor._codeMirror.getGutterElement());
-            gutter.css("min-width", "");
-            var curWidth = gutter.width();
-            if (curWidth > maxWidth) {
-                maxWidth = curWidth;
-            }
-        });
-        
-        if (editors.length === 1) {
-            //There's only the host, just bail
-            editors[0]._codeMirror.setOption("gutter", true);
-            return;
-        }
-        
-        maxWidth = maxWidth + "px";
-        editors.forEach(function (editor) {
-            $(editor._codeMirror.getGutterElement()).css("min-width", maxWidth);
-            editor._codeMirror.setOption("gutter", true);
-        });
-    };
-
-    // Called any time inline was closed, whether manually (via closeThisInline()) or automatically
-    InlineEditor.prototype.onClosed = function () {
-        this.syncGutterWidths();
-        this.editor.destroy(); //release ref on Document
-    };
-
-    // Some tasks have to wait until we've been parented into the outer editor
-    InlineEditor.prototype.onAdded = function (inlineId) {
-        this.inlineId = inlineId;
-            
-        this.syncGutterWidths();
-        
-        // Set initial size
-        this.sizeInlineEditorToContents();
-        
-        this.editor.focus();
-    };
-    InlineEditor.prototype.load = function () {};
-
-    // Called when the editor containing the inline is made visible.
-    InlineEditor.prototype.afterParentShown = function () {
-            sizeInlineEditorToContents(true);
-    };
-
-    // Update the inline editor's height when the number of lines change
-    InlineEditor.prototype.sizeInlineEditorToContents = function (force) {
-        if ($(this.editor._codeMirror.getWrapperElement()).is(":visible")) {
-            var height = this.editor.totalHeight(true);
-            if (force || height !== this.prevHeight) {
-                this.prevHeight = height;
-                this.hostEditor.setInlineWidgetHeight(this.inlineId, height, true);
-                $(this.editor.getScrollerElement()).height(height);
-                this.editor.refresh();
-            }
-        }
-    }
-
-    // TODO (jasonsj): global command
-    // Used to manually trigger closing this inline
-    InlineEditor.prototype.closeEditor = function (inlineEditor) {
-        var shouldMoveFocus = this.editor.hasFocus();
-        EditorManager.closeInlineWidget(this.hostEditor, this.inlineId, shouldMoveFocus);
-        // closeInlineWidget() causes afterClosed() to get run
-    }
-    
-
-    /**
-     * @constructor
-     *
-     */
-    function CSSInlineEditor(rules) {
-        this._rules = rules;
-    }
-    CSSInlineEditor.prototype = new InlineEditor();
-    CSSInlineEditor.prototype.constructor = CSSInlineEditor;
-    CSSInlineEditor.prototype.parentClass = InlineEditor;
-    
-    // TY TODO: rename load?
-    /** @param {!Editor} hostEditor  Outer Editor instance that inline editor will sit within.
-    */
-    CSSInlineEditor.prototype.load = function (hostEditor) {
-        this.hostEditor = hostEditor;
-        var self = this;
-        var htmlContent = document.createElement("div"),
-            ruleList = $("<ul class='pills pills-vertical pull-right'/>");
-        
-        // create rule list
-        this._rules.forEach(function (rule) {
-            ruleList.append("<li><a>" + rule.document.file.name + "</a></li>");
-        });
-        
-        // load first rule
-        var rule = this._rules[0];
-        
-        // create an editor for the first rule
-        var range = {
-            startLine: rule.lineStart,
-            endLine: rule.lineEnd
-        };
-        var inlineInfo = EditorManager.createInlineEditorForDocument(rule.document, range, function(inlineEditor) {
-            self.closeEditor(inlineEditor);
-        });
-
-        // TY TODO: better way to do this than return two types and assign?
-        this.htmlContent = inlineInfo.content;
-        this.editor = inlineInfo.editor;
-
-        var $htmlContent = $(htmlContent);
-        $htmlContent.append(inlineInfo.content);
-        $htmlContent.append(ruleList);
-        
-        // wrapper div for inline editor
-        this.htmlContent = htmlContent;
-
-        // TY TODO: this probably belongs on Editor, but it needs to assigned after this.editor is valid
-        // When text is edited, auto-resize UI and sync changes to a backing full-size editor
-        $(this.editor).on("change", function () {
-            self.sizeInlineEditorToContents();
-        });
-        
-        // TODO (jasonsj): XD
-        _createInlineEditorDecorations(inlineInfo.editor, rule.document);
-        
-        _htmlToCSSProviderContent.push(inlineInfo.content);
-        
-        // Manually position filename div's. Can't use CSS positioning in this case
-        // since the label is relative to the window boundary, not CodeMirror.
-        if (_htmlToCSSProviderContent.length > 0) {
-            $(window).on("resize", _updateAllFilenames);
-            $(DocumentManager).on("dirtyFlagChange", _dirtyFlagChangeHandler);
-        }
-        
-        return (new $.Deferred()).resolve();
-    };
-
-    CSSInlineEditor.prototype.onClosed = function () {
-        this.parentClass.prototype.onClosed.call(this); // call super.onClosed()
-        _inlineEditorRemoved(this.content);
-    }
-    
-    CSSInlineEditor.prototype.getRules = function () {
-    };
-    
-    CSSInlineEditor.prototype.getSelectedRule = function () {
-    };
-    
-    CSSInlineEditor.prototype.setSelectedRule = function () {
-    };
-    
-    CSSInlineEditor.prototype.nextRule = function () {
-    };
-    
-    CSSInlineEditor.prototype.previousRule = function () {
-    };
-    
-    function FindInFilesInline() {
-    }
-    FindInFilesInline.prototype = new InlineEditor();
     
     /**
      * When cursor is on an HTML tag name, class attribute, or id attribute, find associated
@@ -399,7 +212,7 @@ define(function (require, exports, module) {
         CSSUtils.findMatchingRules(selectorName)
             .done(function (rules) {
                 if (rules && rules.length > 0) {
-                    var cssInlineEditor = new CSSInlineEditor(rules);
+                    var cssInlineEditor = new CSSInlineEditor.CSSInlineEditor(rules);
                     
                     cssInlineEditor.load(hostEditor).done(function () {
                         result.resolve(cssInlineEditor);
