+++ test/spec/EditorManager-test.js
@@ -70,13 +70,13 @@ define(function (require, exports, module) {
             });
             
             it("should not add a doc to the working set if no changes where made", function () {
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 var i = DocumentManager.findInWorkingSet(this.path + "/test.css");
                 expect(i).toEqual(-1);
             });
             
             it("should add a new doc to the working set and sync it with the inline text when a change is made", function () {
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 var i = DocumentManager.findInWorkingSet(this.path + "/test.css");
                 expect(i).toEqual(-1);
                 
@@ -89,13 +89,13 @@ define(function (require, exports, module) {
             it("should use an already open doc and sync with it from the inline text when a change is made", function () {
                 DocumentManager.addToWorkingSet(cssDoc);
                 
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 inlineInfo.editor._setText(newCss);
                 expect(cssDoc.getText()).toEqual(FileUtils.translateLineEndings(newCss, cssDoc._lineEndings));
             });
             
             it("should sync even if the contents of the inline are all deleted", function () {
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 
                 inlineInfo.editor._setText("");
                 expect(cssDoc.getText()).toEqual("");
@@ -103,7 +103,7 @@ define(function (require, exports, module) {
             
             it("should sync after an undoing and redoing an edit", function () {
                 var oldCss = cssDoc.getText();
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 
                 inlineInfo.editor._setText(newCss);
                 expect(cssDoc.getText()).toEqual(FileUtils.translateLineEndings(newCss, cssDoc._lineEndings));
@@ -116,7 +116,7 @@ define(function (require, exports, module) {
             });
             
             it("should sync multiple edits in the inline", function () {
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 
                 var curText = "";
                 var i = 0;
@@ -131,7 +131,7 @@ define(function (require, exports, module) {
             /* we don't currently support syncing from the other direction, so when that gets
                added, then this unit test should get reversed and more added for that story */
             it("should *NOT* sync changes from the main document back to the inline editor", function () {
-                var inlineInfo = EditorManager.createInlineEditorForDocument(myEditor, cssDoc, cssDocRange);
+                var inlineInfo = EditorManager.createInlineEditorForDocument(cssDoc, cssDocRange);
                 
                 inlineInfo.editor._setText(newCss);
                 newCss = "h1 {\n    background-color: #0F0;\n}";
